TRAC-8 Refactor GroupSizeViews
- add new view GroupSizeView
- extend this new class in the GroupEraseSizeView and GroupPencilSizeView
